mithril_cli_helper::serde_deserialization

Function string_or_struct

pub fn string_or_struct<'de, T, D>(deserializer: D) -> Result<T, D::Error>
where
    T: Deserialize<'de> + FromStr,
    D: Deserializer<'de>,
    T::Err: Display,

Deserialize a value that can be either a string or a struct if it implements FromStr.

Useful for deserializing environment variables that can be passed as a JSON string or as a map. Ie: for the following enum:

#[derive(Debug, serde::Deserialize)]
#[serde(tag = "type", rename_all = "lowercase")]
enum EnumParam {
   Foo { foo: String },
}

impl FromStr for EnumParam {
  type Err = serde_json::Error;

  fn from_str(s: &str) -> Result<Self, Self::Err> {
      serde_json::from_str(s)
  }
}

this allows config to deserialize the following environment variables:

export ENUM_PARAM='{"type":"foo", "foo": "foo value"}'

or

export ENUM_PARAM__TYPE="foo"
export ENUM_PARAM__FOO="foo value"

Code based on: https://serde.rs/string-or-struct.html