mithril_cli_helper::serde_deserialization

Function string_or_struct_optional

pub fn string_or_struct_optional<'de, T, D>(
    deserializer: D,
) -> Result<Option<T>, D::Error>
where
    T: Deserialize<'de> + FromStr + 'de,
    D: Deserializer<'de>,
    T::Err: Display,

Deserialize an optional value that can be either a string or a struct if it implements FromStr.

Important: to handle the case where the key is missing in the JSON file, the #[serde(default)] attribute must be set.

Useful for deserializing environment variables that can be passed as a JSON string or as a map. Ie: for the following enum:

use mithril_cli_helper::serde_deserialization::string_or_struct_optional;

#[derive(Debug, PartialEq, serde::Deserialize)]
#[serde(tag = "type", rename_all = "lowercase")]
enum EnumParam {
   Foo { foo: String },
}

impl FromStr for EnumParam {
  type Err = serde_json::Error;

  fn from_str(s: &str) -> Result<Self, Self::Err> {
      serde_json::from_str(s)
  }
}

#[derive(Debug, PartialEq, serde::Deserialize)]
struct Config {
    #[serde(default, deserialize_with = "string_or_struct_optional")]
    enum_param: Option<EnumParam>,
}

// Deserialize from null
let config: Config = serde_json::from_str(r#"{"enum_param": null}"#).unwrap();
assert_eq!(config.enum_param, None);

// Deserialize from missing key
let config: Config = serde_json::from_str(r#"{}"#).unwrap();
assert_eq!(config.enum_param, None);

// Deserialize from object
let config: Config = serde_json::from_str(r#"{"enum_param": {"type":"foo", "foo": "bar"}}"#).unwrap();
assert_eq!(config.enum_param, Some(EnumParam::Foo { foo: "bar".to_string() }));

this allows config to deserialize the following environment variables:

export ENUM_PARAM='{"type":"foo", "foo": "foo value"}'

or

export ENUM_PARAM__TYPE="foo"
export ENUM_PARAM__FOO="foo value"

Code based on: https://serde.rs/string-or-struct.html